ksy974498/app/pyproject.toml
[tool.poetry]
name = "sigenie"
version = "0.1.0"
description = ""
authors = ["Kwang Yong Jung <gyjong@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<=3.13"
pandas = "^2.2.2"
langchain = "^0.2.6"
langchain-openai = "^0.1.14"
python-dotenv = "^1.0.1"
langgraph = "^0.1.5"
langchain-groq = "^0.1.6"
google-generativeai = "^0.7.1"
tabulate = "^0.9.0"
streamlit = "^1.36.0"
langchain-anthropic = "^0.1.20"
langchain-google-genai = "^1.0.7"
langchain-experimental = "^0.0.62"
bs4 = "^0.0.2"
pypdf = "^4.3.0"
faiss-cpu = "^1.8.0.post1"
duckduckpy = "^0.2"
duckduckgo-search = "^6.2.1"
#autogen = "*"
crewai = "^0.41.1"
#pyaudio = "^0.2.14"
speechrecognition = "^3.10.4"
gtts = "^2.5.2"
pygame = "^2.6.0"
pyttsx3 = "^2.90"
matplotlib = "^3.9.1"
networkx = "^3.3"
neo4j = "^5.23.1"
graphviz = "^0.20.3"
pyvis = "^0.3.2"
pymongo = {version = "3.11", extras = ["srv"]}



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
